backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
httpx==0.27.2
openai==1.55.3
pytest==8.3.3
pytest-asyncio==0.24.0
email-validator==2.2.0
python-multipart==0.0.9
